can load a single pdf into the project
