--- ReaderAndroid/app/src/main/java/com/example/readerandroid/Cover.kt
@@ -1,8 +1,0 @@
-package com.kontinua.readerandroid
-
-data class Cover(
-    val id: String,
-    val desc: String,
-    val videos: List<Video>?,
-    val references: List<Reference>?
-)
