--- ReaderAndroid/app/src/main/java/com/example/readerandroid/Reference.kt
@@ -1,8 +1,0 @@
-package com.kontinua.readerandroid
-
-
-data class Reference(
-    val id: String,
-    val link: String,
-    val title: String
-)
