--- ReaderAndroid/app/src/main/java/com/example/readerandroid/RetrofitClient.kt
@@ -1,28 +1,0 @@
-package com.kontinua.readerandroid
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-
-object RetrofitClient {
-
-    private const val BASE_URL = "http://localhost:8000/" // Your server URL
-
-    // Optional: Add logging interceptor for debugging
-    private val loggingInterceptor = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY // Or .BASIC, .HEADERS, etc.
-    }
-
-    private val okHttpClient = OkHttpClient.Builder()
-        .addInterceptor(loggingInterceptor)
-        .build()
-
-    private val retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .client(okHttpClient) // Add the logging interceptor
-        .build()
-
-    val apiService: ApiService = retrofit.create(ApiService::class.java)
-}
