--- ReaderAndroid/app/src/main/java/com/example/readerandroid/Video.kt
@@ -1,7 +1,0 @@
-package com.kontinua.readerandroid
-
-data class Video(
-    val id: String,
-    val link: String,
-    val title: String
-)
